+++ members/nullnet-server/src/db/migrations/2026-09-15-000001_observations/down.sql
@@ -1,0 +1,4 @@
+DROP TRIGGER observation_backend_session;
+DROP TRIGGER observation_seed;
+DROP TABLE observation_counts;
+DROP TABLE observations;